requirements.txt
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
JayDeBeApi==1.1.1
Jinja2==2.10.1
JPype1==0.6.3
MarkupSafe==1.1.1
pytz==2019.2
six==1.12.0
uWSGI==2.0.18
Werkzeug==0.15.5
